+++ dicom_validator/tests/validator/test_iod_validator_func_groups.py
@@ -82,7 +82,6 @@ FRAME_ANATOMY = {
                 {
                     "CodeValue": "T-D3000",
                     "CodingSchemeDesignator": "SRT",
-                    "CodingSchemeVersion": "1",
                     "CodeMeaning": "Chest",
                 }
             ],