+++ CHANGES.md
@@ -6,6 +6,9 @@ The released versions correspond to PyPi releases.
 ### Features
 * added initial support for validating functional group macros
 
+### Fixes
+* fixes handling of unverifiable and condition, see [#32](../.. /issues/32)
+
 ### Infrastructure
 * Added pre-commit configuration for use with pre-commit hook